+++ blocks/mai-popup/block.php
@@ -34,15 +34,18 @@ function mai_register_popup_block() {
  * @return void
  */
 function mai_do_popup_block( $attributes, $content, $is_preview, $post_id, $wp_block, $context ) {
-	      $args           = [];
+	$args                 = [];
 	$args['class']        = isset( $attributes['className'] ) ? $attributes['className']: '';
 	$args['position']     = isset( $attributes['alignContent'] ) ? $attributes['alignContent']: '';
+	$args['background']   = isset( $attributes['backgroundColor'] ) ? $attributes['backgroundColor'] : '';
+	$args['color']        = isset( $attributes['textColor'] ) ? $attributes['textColor'] : '';
 	$args['id']           = get_field( 'id' );
 	$args['trigger']      = get_field( 'trigger' );
 	$args['animate']      = get_field( 'animate' );
 	$args['text']         = get_field( 'text' );
 	$args['delay']        = get_field( 'delay' );
 	$args['width']        = get_field( 'width' );
+	$args['padding']      = get_field( 'padding' );
 	$args['repeat']       = get_field( 'repeat' );
 	$args['repeat_roles'] = get_field( 'repeat_roles' );
 	$args['preview']      = $is_preview;
@@ -68,17 +71,20 @@ function mai_register_popup_field_group() {
 		return;
 	}
 
+	$defaults = maipopups_get_defaults();
+
 	acf_add_local_field_group(
 		[
 			'key'    => 'mai_popup_field_group',
 			'title'  => __( 'Mai Popup', 'mai-popups'),
 			'fields' => [
 				[
-					'key'     => 'mai_popup_trigger',
-					'label'   => __( 'Trigger', 'mai-popups' ),
-					'name'    => 'trigger',
-					'type'    => 'select',
-					'choices' => [
+					'key'           => 'mai_popup_trigger',
+					'label'         => __( 'Trigger', 'mai-popups' ),
+					'name'          => 'trigger',
+					'type'          => 'select',
+					'default_value' => $defaults['trigger'],
+					'choices'       => [
 						'manual' => __( 'Manual (Custom Link)', 'mai-popups' ),
 						'load'   => __( 'On Load', 'mai-popups' ),
 						'scroll' => __( 'Scroll Distance', 'mai-popups' ),
@@ -86,11 +92,12 @@ function mai_register_popup_field_group() {
 					],
 				],
 				[
-					'key'     => 'mai_popup_animate',
-					'label'   => __( 'Animation', 'mai-popups' ),
-					'name'    => 'animate',
-					'type'    => 'select',
-					'choices' => [
+					'key'           => 'mai_popup_animate',
+					'label'         => __( 'Animation', 'mai-popups' ),
+					'name'          => 'animate',
+					'type'          => 'select',
+					'default_value' => $defaults['animate'],
+					'choices'       => [
 						'fade' => __( 'Fade In', 'mai-popups' ),
 						'up'   => __( 'Slide up', 'mai-popups' ),
 						'down' => __( 'Slide down', 'mai-popups' ),
@@ -101,7 +108,7 @@ function mai_register_popup_field_group() {
 					'label'             => __( 'Scroll distance', 'mai-popups' ),
 					'name'              => 'distance',
 					'type'              => 'number',
-					'default_value'     => 50,
+					'default_value'     => $defaults['distance'],
 					'min'               => 0,
 					'max'               => '',
 					'step'              => 1,
@@ -121,6 +128,7 @@ function mai_register_popup_field_group() {
 					'label'             => __( 'Delay', 'mai-popups' ),
 					'name'              => 'delay',
 					'type'              => 'number',
+					'default_value'     => $defaults['delay'],
 					'min'               => 0,
 					'max'               => '',
 					'step'              => '.5',
@@ -136,12 +144,29 @@ function mai_register_popup_field_group() {
 					],
 				],
 				[
-					'key'          => 'mai_popup_width',
-					'label'        => __( 'Width', 'mai-popups' ),
-					'instructions' => __( 'Accepts any CSS value (px, em, rem, vw, ch, etc.). Using 100% removes margin around content.', 'mai-popups' ),
-					'name'         => 'width',
-					'type'         => 'text',
-					'placeholder'  => '600px',
+					'key'           => 'mai_popup_width',
+					'label'         => __( 'Width', 'mai-popups' ),
+					'instructions'  => __( 'Accepts any CSS value (px, em, rem, vw, ch, etc.). Using 100% removes margin around content.', 'mai-popups' ),
+					'name'          => 'width',
+					'type'          => 'text',
+					'placeholder'   => '600px',
+					'default_value' => $defaults['width'],
+				],
+				[
+					'key'           => 'mai_popup_padding',
+					'label'         => __( 'Padding', 'mai-popups' ),
+					'name'          => 'padding',
+					'type'          => 'select',
+					'default_value' => $defaults['padding'],
+					'choices'       => [
+						''     => __( 'None', 'mai-popups' ),
+						'sm'   => 'SM',
+						'md'   => 'MD',
+						'lg'   => 'LG',
+						'xl'   => 'XL',
+						'xxl'  => 'XXL',
+						'xxxl' => 'XXXL',
+					]
 				],
 				[
 					'key'               => 'mai_popup_repeat',
@@ -149,7 +174,7 @@ function mai_register_popup_field_group() {
 					'instructions'      => __( 'The time it takes before this popup will be displayed again for the same user. Use 0 to always show, but beware that this may frustrate your website users.', 'mai-popups' ),
 					'name'              => 'repeat',
 					'type'              => 'text',
-					'default_value'     => '7 days', // Can't translate. English for `strtotime()`.
+					'default_value'     => $defaults['repeat'],
 					'conditional_logic' => [
 						[
 							[
@@ -166,6 +191,7 @@ function mai_register_popup_field_group() {
 					'name'              => 'repeat_roles',
 					'instructions'      => __( 'Select user roles that will always see this popup, regardless of the setting above.', 'mai-popups' ),
 					'type'              => 'select',
+					'default_value'     => $defaults['repeat_roles'],
 					'choices'           => [], // Added later.
 					'return_format'     => 'value',
 					'multiple'          => 1,
@@ -202,8 +228,38 @@ function mai_register_popup_field_group() {
 	);
 }
 
+add_filter( 'acf/pre_render_field', 'mai_pre_render_popup_padding_field', 10, 2 );
+/**
+ * Unsets default value if this is an instance of the block
+ * added before the padding field was available.
+ *
+ * @since TBD
+ *
+ * @param array $field
+ * @param mixed $post_id
+ *
+ * @return array
+ */
+function mai_pre_render_popup_padding_field( $field, $post_id ) {
+	// Bail if not the field we want.
+	if ( 'mai_popup_padding' !== $field['key'] ) {
+		return $field;
+	}
+
+	// Get all field values.
+	$fields = get_fields();
+
+	// If no fields were saved, this is a new instance of the block.
+	if ( $fields && ! isset( $fields['padding'] ) ) {
+		$field['default_value'] = '';
+	}
+
+	return $field;
+}
+
 add_filter( 'acf/load_field/key=mai_popup_repeat_roles', 'mai_load_popup_repeat_roles' );
 /**
+ * Loads existing roles as field choices.
  *
  * @since 0.4.0
  *